- Deleted CSVFile entity (not useful) - Finished implementation of CSVFileProcessing - Fixed ReadableFile interface (i.e. implemented ReadRows properly) - Finished implementation of SongPool entity
