--- src/main/java/Entities/CSVFile.java
@@ -1,10 +1,0 @@
-package Entities;
-
-import java.util.ArrayList;
-
-public class CSVFile implements ReadableFile {
-    @Override
-    public ArrayList<String> readFile() {
-        return null;
-    }
-}
