+++ src/main/java/Entities/ReadableFile.java
@@ -7,5 +7,5 @@ public interface ReadableFile {
      * Read the file type specified
      **/
 
-     ArrayList<String> readFile(); // Returns an array list containing String for each line of file (i.e. csv, etc.)
+     ArrayList<String> readRows(String filePath); // Returns an array list containing String for each line of file (i.e. csv, etc.)
 }
