+++ src/main/java/Entities/SongPool.java
@@ -1,5 +1,4 @@
 package Entities;
-import Processors.CSVFileProcessing;
 
 import java.util.ArrayList;
 
@@ -10,12 +9,29 @@ public class SongPool {
         this.songs = new ArrayList<>();
     }
 
-    public void populateSongPool(ReadableFile songFile) {
-        ArrayList<String> songsReadFromFile = songFile.readFile(); //want readFile to return a list of Songs, we don't care how
+    public void populateSongPool(ReadableFile songFile, String filePath) {
+        ArrayList<String> songsReadFromFile = songFile.readRows(filePath); //want readFile to return a list of Songs, we don't care how
         for (String readLine : songsReadFromFile) {
             String[] values = readLine.split(",");
-            this.songs.add(CSVFileProcessing.readLine(values));
+            this.songs.add(readLine(values));
         }
+    }
+
+    // helper function
+    public static Song readLine(String[] values) {
+        Song newSong = new Song();
+
+        newSong.song = values[0];
+        newSong.artist = values[1];
+        newSong.genre = values[2];
+        newSong.energy = values[5];
+        newSong.liveness = values[8];
+        newSong.valence = values[9];
+        newSong.popularity = values[13];
+        newSong.danceability = values[6];
+        newSong.bpm = values[4];
+
+        return newSong;
     }
 
 }
