+++ src/main/java/Processors/CSVFileProcessing.java
@@ -1,7 +1,6 @@
 package Processors;
 
-import Entities.CSVFile;
-import Entities.Song;
+import Entities.ReadableFile;
 
 import java.io.BufferedReader;
 import java.io.FileNotFoundException;
@@ -9,47 +8,31 @@ import java.io.FileReader;
 import java.io.IOException;
 import java.util.ArrayList;
 
-public class CSVFileProcessing {
+public class CSVFileProcessing implements ReadableFile {
 
     /**
      *
-     * @param values containing processed CSV line (i.e. a string representing the line read from CSV data set we chose)
-     * @return Song object based on the read line
+     *
+     * @param filePath that represents the directory/path of the dataset calculated on
+     * @return an array list of strings where each element/string represents a line of data in the data set (CSV)
      */
-    public static Song readLine(String[] values) {
-        Song newSong = new Song();
-
-        newSong.song = values[0];
-        newSong.artist = values[1];
-        newSong.genre = values[2];
-        newSong.energy = values[5];
-        newSong.liveness = values[8];
-        newSong.valence = values[9];
-        newSong.popularity = values[13];
-        newSong.danceability = values[6];
-        newSong.bpm = values[4];
-
-        return newSong;
-    }
-
-
-    public static void main(String[] args) {
-        String path = "/Users/alex/Desktop/SacramentocrimeJanuary2006.csv";
+    @Override
+    public ArrayList<String> readRows(String filePath) {
         String line = "";
-        ArrayList<Song> songs = new ArrayList<>();
+        ArrayList<String> rowsRead = new ArrayList<>();
 
         try {
-            BufferedReader br = new BufferedReader(new FileReader(path));
+            BufferedReader br = new BufferedReader(new FileReader(filePath));
 
             while((line = br.readLine()) != null) {
-                String[] values = line.split(",");
-                songs.add(readLine(values));
+                rowsRead.add(line);
             }
         } catch (FileNotFoundException error) {
             error.printStackTrace();
         } catch (IOException error) {
             error.printStackTrace();
         }
-    }
 
+        return rowsRead;
+    }
 }
